requirements.txt
lavaplayer==1.0.10a0
aiohttp==3.8.1
frozenlist==1.3.1
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.0
multidict==6.0.2
aioredis==2.0.1
typing_extensions==4.3.0
hikari-lightbulb==2.2.4
hikari==2.0.0.dev110
aiosignal==1.2.0
yarl==1.8.1
idna==3.3
colorlog==6.6.0
colorama==0.4.5
toml==0.10.2